refactor(sharing-dialog): rename sharing-constants to constants
