+++ components/sharing-dialog/src/access-add/access-add.js
@@ -5,9 +5,9 @@ import { colors } from '@dhis2/ui-constants'
 import PropTypes from 'prop-types'
 import React, { useState, useContext } from 'react'
 import { SharingAutocomplete } from '../autocomplete/index.js'
+import { ACCESS_VIEW_ONLY, ACCESS_VIEW_AND_EDIT } from '../constants.js'
 import { FetchingContext } from '../context/index.js'
 import i18n from '../locales/index.js'
-import { ACCESS_VIEW_ONLY, ACCESS_VIEW_AND_EDIT } from '../sharing-constants.js'
 import { Title } from '../text/index.js'
 
 export const AccessAdd = ({ onAdd }) => {
