+++ components/sharing-dialog/src/constants.js
